Changed the way the db file was backed up
Previously:
the db file was moved into the backup folder
and new db file was written to replace the one moved.

Now:
The db file is copied into the backup folder
The new existing db file has it’s content replaced with the new content.

Why:
Some file syncing services found the ‘move, then create a new file with the same name as the old one’ confusing and could end up with sync conflicts that needed to be resolved if syncing to multiple machines. The new approach creates a new file and just updates the content of the existing file, so you don’t experience 2 different files with the same name sometimes.

Anyway, syncing services appear more stable with this new approach.